processImage in xmippMetadataProgram now passes rowIn and rowOut
git-svn-id: file:///home/roberto/Git/svn@10285 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 0cbba8125874872a405a17a97a808ef00e54d023

+++ trunk/xmipp/libraries/data/metadata_label.cpp
@@ -746,7 +746,7 @@ void MDRow::copy(const MDRow &row)
         }
     }
     //copy the order of labels
-    memcpy(order, row.order, sizeof(short int)*_size);
+    memcpy(order, row.order, sizeof(int)*_size);
 }
 
 std::ostream& operator << (std::ostream &out, const MDRow &row)
